requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-cache2==0.2.1
filelock==3.9.0
fsspec==2023.4.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
orjson==3.9.15
pendulum==3.0.0
pillow==10.2.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==4.6.0
scikit-learn==1.2.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
threadpoolctl==3.4.0
time-machine==2.14.0
typing_extensions==4.8.0
tzdata==2024.1
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0

gunicorn
